+++ CHANGES.md
@@ -18,6 +18,7 @@ Changes in 7.0
 - Merge Average Daily Attendance & Average Attendance by Day in Menu.php, Percent.php, Help_en.php, help.po & rosariosis.sql
 - Remove "Happy []..." text in Portal.php
 - HTML remove "Demographics" header to gain space on PDF in AttendanceSummary.php
+- SQL Update ATTENDANCE_CODE (admin) when is NULL in TakeAttendance.php
 
 Changes in 6.9.4
 ----------------